feat: 새로고침했을때 api 요청 설정
