+++ client/src/App.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { useRecoilValue } from "recoil";
 import { userState } from "@recoil/user-state";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
@@ -12,51 +12,57 @@ import StudyRoom from "@pages/StudyRoom/StudyRoom";
 import NotFoundPage from "@pages/NotFoundPage/NotFoundPage";
 import RedirectPage from "@pages/Login/RedirectPage";
 import { getNewToken } from "@api/auth-api";
+import { Loading } from "@components/commons";
 
 function App() {
   const user = useRecoilValue(userState);
+  const [isLoading, setIsLoading] = useState(false);
 
   // 새로고침했을 때도 로그인 상태 계속 유지
   useEffect(async () => {
-    console.log("로그인?", user.isLogin); // 확인용
     if (!user.isLogin) {
       return;
     }
+    setIsLoading((prev) => !prev);
     await getNewToken();
+    setIsLoading((prev) => !prev);
   }, []);
 
   return (
-    <BrowserRouter>
-      <Switch>
-        <Route exact path="/">
-          <MainHome />
-        </Route>
-        <Route exact path="/login">
-          <Login />
-        </Route>
-        <Route path="/login/oauth2/code/kakao">
-          <RedirectPage />
-        </Route>
-        <Route path="/search">
-          <Search />
-        </Route>
-        <Route path="/mypage">
-          <MyPage />
-        </Route>
-        <Route path="/studyroom/create">
-          <CreateRoom />
-        </Route>
-        <Route exact path="/studyroom/:id">
-          <StudyRoom />
-        </Route>
-        <Route path="/studyroom/:id/setting">
-          <JoinRoom />
-        </Route>
-        <Route path="*">
-          <NotFoundPage />
-        </Route>
-      </Switch>
-    </BrowserRouter>
+    <>
+      {isLoading && <Loading />}
+      <BrowserRouter>
+        <Switch>
+          <Route exact path="/">
+            <MainHome />
+          </Route>
+          <Route exact path="/login">
+            <Login />
+          </Route>
+          <Route path="/login/oauth2/code/kakao">
+            <RedirectPage />
+          </Route>
+          <Route path="/search">
+            <Search />
+          </Route>
+          <Route path="/mypage">
+            <MyPage />
+          </Route>
+          <Route path="/studyroom/create">
+            <CreateRoom />
+          </Route>
+          <Route exact path="/studyroom/:id">
+            <StudyRoom />
+          </Route>
+          <Route path="/studyroom/:id/setting">
+            <JoinRoom />
+          </Route>
+          <Route path="*">
+            <NotFoundPage />
+          </Route>
+        </Switch>
+      </BrowserRouter>
+    </>
   );
 }
