+++ client/src/components/home/Header/Header.jsx
@@ -1,15 +1,27 @@
-import React, { useState } from "react";
-import { useRecoilValue } from "recoil";
+import React, { useState, useEffect } from "react";
+import axiosInstance from "@api/axios-config";
+import { useRecoilState } from "recoil";
 import { useHistory } from "react-router-dom";
 import { userState } from "@recoil/user-state";
 import { logout } from "@api/auth-api";
 import { Search, Profile } from "@icons";
+import { getNickname } from "@api/user-api";
 import styles from "./Header.module.css";
 
 function Header() {
   const history = useHistory();
-  const user = useRecoilValue(userState);
   const [isDropdown, setIsDropdown] = useState(false);
+  const [user, setUserState] = useRecoilState(userState);
+  const axiosAuthorization = axiosInstance.defaults.headers.common["Authorization"];
+
+  useEffect(async () => {
+    if (!user.isLogin || user.nickname !== null || !axiosAuthorization) {
+      return;
+    }
+    await getNickname()
+      .then((response) => setUserState((prev) => ({ ...prev, nickname: response.nickname })))
+      .catch((error) => console.error("get nickname error", error));
+  }, [axiosAuthorization]);
 
   const ClickStudyRoom = () => {
     history.push({
