+++ client/src/pages/MyPage/MyPage.jsx
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from "react";
+import axiosInstance from "@api/axios-config";
 import { useRecoilValue } from "recoil";
 import { userState } from "@recoil/user-state";
 import { getProfile, getTimeRecordList, getMyRoom } from "@api/mypage-api";
@@ -22,14 +23,16 @@ import getTimeDiff from "src/utils/getTimeDiff";
 import styles from "./MyPage.module.css";
 
 function MyPage() {
+  const axiosAuthorization = axiosInstance.defaults.headers.common["Authorization"];
+
   const {
     data: profileData, //
     setData: updateProfileData,
-  } = useAsync(getProfile, { onError: (error) => console.error(error) }, [], false);
+  } = useAsync(getProfile, { onError: (error) => console.error(error) }, [axiosAuthorization], false);
   const {
     data: myRoomData, //
     setData: updateRoomData,
-  } = useAsync(getMyRoom, { onError: (error) => console.error(error) }, [], false);
+  } = useAsync(getMyRoom, { onError: (error) => console.error(error) }, [axiosAuthorization], false);
   const [selectedDate, setSelectedDate] = useState(new Date().toISOString().slice(0, 10));
   const [timeRecordData, setTimeRecordData] = useState();
   const [totalStudyTime, setTotalStudyTime] = useState();
@@ -57,11 +60,14 @@ function MyPage() {
 
   useEffect(() => {
     try {
+      if (!axiosAuthorization) {
+        return;
+      }
       fetchTimeRecordData(selectedDate);
     } catch (error) {
       console.error(error);
     }
-  }, [selectedDate]);
+  }, [selectedDate, axiosAuthorization]);
 
   const renderEditModal = (type) => setIsModalOpen(type);
